CEDS-4840: consider non-amendment on removing items to show either content
