+++ app/controllers/declaration/ItemsSummaryController.scala
@@ -16,25 +16,22 @@
 
 package controllers.declaration
 
-import connectors.CustomsDeclareExportsConnector
 import controllers.actions.{AuthAction, JourneyAction}
 import controllers.declaration.routes.{ItemsSummaryController, ProcedureCodesController, TransportLeavingTheBorderController}
 import controllers.helpers.SequenceIdHelper
 import controllers.navigation.Navigator
 import forms.common.YesNoAnswer
 import forms.common.YesNoAnswer.YesNoAnswers
-import handlers.ErrorHandler
-import models.DeclarationType.CLEARANCE
 import models.ExportsDeclaration
 import models.declaration.ExportItem
 import models.requests.JourneyRequest
 import play.api.data.{Form, FormError}
 import play.api.i18n.I18nSupport
-import play.api.mvc.{Action, AnyContent, MessagesControllerComponents, Result}
+import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import services.cache.{ExportItemIdGeneratorService, ExportsCacheService}
 import uk.gov.hmrc.play.bootstrap.controller.WithUnsafeDefaultFormBinding
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
-import views.html.declaration.declarationitems.{items_add_item, items_cannot_remove, items_remove_item, items_summary}
+import views.html.declaration.declarationitems.{items_add_item, items_summary}
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
@@ -43,21 +40,16 @@ class ItemsSummaryController @Inject() (
   authenticate: AuthAction,
   journeyType: JourneyAction,
   exportsCacheService: ExportsCacheService,
-  customsDeclareExportsConnector: CustomsDeclareExportsConnector,
   navigator: Navigator,
-  errorHandler: ErrorHandler,
   exportItemIdGeneratorService: ExportItemIdGeneratorService,
   mcc: MessagesControllerComponents,
   addItemPage: items_add_item,
   itemsSummaryPage: items_summary,
-  itemsCannotRemovePage: items_cannot_remove,
-  removeItemPage: items_remove_item,
   sequenceIdHandler: SequenceIdHelper
 )(implicit ec: ExecutionContext)
     extends FrontendController(mcc) with I18nSupport with WithUnsafeDefaultFormBinding {
 
   private def itemSummaryForm: Form[YesNoAnswer] = YesNoAnswer.form(errorKey = "declaration.itemsSummary.addAnotherItem.error.empty")
-  private def removeItemForm: Form[YesNoAnswer] = YesNoAnswer.form(errorKey = "declaration.itemsRemove.error.empty")
 
   def displayAddItemPage(): Action[AnyContent] = (authenticate andThen journeyType) { implicit request =>
     if (request.cacheModel.items.isEmpty) Ok(addItemPage())
@@ -118,80 +110,4 @@ class ItemsSummaryController @Inject() (
     exportsCacheService.update(request.cacheModel.copy(items = itemsSequenced, declarationMeta = updatedMeta))
   }
 
-  def displayRemoveItemConfirmationPage(itemId: String, fromSummary: Boolean = false): Action[AnyContent] = (authenticate andThen journeyType).async {
-    implicit request =>
-      request.cacheModel.itemWithIndexBy(itemId) match {
-        case Some((item, idx)) =>
-          findParentDeclaration flatMap {
-            case Some(parentDeclaration) =>
-              Future.successful(
-                canItemBeRemoved(
-                  item,
-                  parentDeclaration,
-                  Ok(removeItemPage(removeItemForm, item, idx, fromSummary)),
-                  Ok(itemsCannotRemovePage(item, idx, parentDeclaration.id))
-                )
-              )
-            case None => errorHandler.internalError(s"Could not find parentDecId from declaration [${request.cacheModel.id}]")
-          }
-        case None =>
-          Future.successful(navigator.continueTo(ItemsSummaryController.displayItemsSummaryPage))
-      }
-  }
-
-  def removeItem(itemId: String, fromSummary: Boolean = false): Action[AnyContent] = (authenticate andThen journeyType).async { implicit request =>
-    removeItemForm
-      .bindFromRequest()
-      .fold(
-        formWithErrors =>
-          request.cacheModel.itemWithIndexBy(itemId) match {
-            case Some((item, idx)) =>
-              findParentDeclaration flatMap {
-                case Some(parentDeclaration) =>
-                  Future.successful(
-                    canItemBeRemoved(
-                      item,
-                      parentDeclaration,
-                      BadRequest(removeItemPage(formWithErrors, item, idx, fromSummary)),
-                      BadRequest(itemsCannotRemovePage(item, idx, parentDeclaration.id))
-                    )
-                  )
-                case None => errorHandler.internalError(s"Could not find parentDecId from declaration [${request.cacheModel.id}]")
-              }
-            case None => errorHandler.internalError(s"Could not find ExportItem with id = [$itemId] to remove")
-          },
-        _.answer match {
-          case YesNoAnswers.yes =>
-            removeItemFromCache(itemId).map(_ => navigator.continueTo(ItemsSummaryController.displayItemsSummaryPage))
-
-          case YesNoAnswers.no =>
-            Future.successful(navigator.continueTo(ItemsSummaryController.displayItemsSummaryPage))
-        }
-      )
-  }
-
-  private def findParentDeclaration(implicit request: JourneyRequest[AnyContent]): Future[Option[ExportsDeclaration]] =
-    request.cacheModel.declarationMeta.parentDeclarationId match {
-      case Some(parentDecId) =>
-        customsDeclareExportsConnector.findDeclaration(parentDecId)
-      case _ => Future.successful(None)
-    }
-
-  private def canItemBeRemoved(item: ExportItem, parentDec: ExportsDeclaration, remove: Result, cannotRemove: Result): Result =
-    if (!parentDec.items.map(_.sequenceId).contains(item.sequenceId)) remove
-    else cannotRemove
-
-  private def removeItemFromCache(itemId: String)(implicit request: JourneyRequest[AnyContent]): Future[ExportsDeclaration] =
-    request.cacheModel.itemBy(itemId) match {
-      case Some(itemToDelete) =>
-        val filteredItems = request.cacheModel.items.filterNot(_.id == itemToDelete.id)
-        val (updatedItems, updatedMeta) = sequenceIdHandler.handleSequencing(filteredItems, request.cacheModel.declarationMeta)
-        val updatedModel = removeWarehouseIdentification(request.cacheModel.copy(items = updatedItems, declarationMeta = updatedMeta))
-        exportsCacheService.update(updatedModel)
-      case None => Future.successful(request.cacheModel)
-    }
-
-  private def removeWarehouseIdentification(declaration: ExportsDeclaration): ExportsDeclaration =
-    if (declaration.isType(CLEARANCE) || declaration.requiresWarehouseId) declaration
-    else declaration.copy(locations = declaration.locations.copy(warehouseIdentification = None))
 }
