+++ test/views/declaration/declarationitems/ItemsSummaryViewSpec.scala
@@ -107,14 +107,14 @@ class ItemsSummaryViewSpec extends UnitViewSpec with ExportsTestHelper with Stub
         rows.get(1).getElementById("item_0--item_type") must containText("1234567890")
         rows.get(1).getElementById("item_0--package_count").text() must be("1")
         rows.get(1).getElementById("item_0--change").getElementsByTag("a").get(0) must haveHref(routes.ProcedureCodesController.displayPage("id1"))
-        rows.get(1).getElementById("item_0--remove").getElementsByTag("a").get(0) must haveHref(routes.ItemsSummaryController.removeItem("id1"))
+        rows.get(1).getElementById("item_0--remove").getElementsByTag("a").get(0) must haveHref(routes.RemoveItemsSummaryController.removeItem("id1"))
 
         rows.get(2).getElementById("item_1--sequence_id") must containText("2")
         rows.get(2).getElementById("item_1--procedure_code") must containText("procedure-code2")
         rows.get(2).getElementById("item_1--item_type") must containText("1234567890")
         rows.get(2).getElementById("item_1--package_count").text() must be("2")
         rows.get(2).getElementById("item_1--change").getElementsByTag("a").get(0) must haveHref(routes.ProcedureCodesController.displayPage("id2"))
-        rows.get(2).getElementById("item_1--remove").getElementsByTag("a").get(0) must haveHref(routes.ItemsSummaryController.removeItem("id2"))
+        rows.get(2).getElementById("item_1--remove").getElementsByTag("a").get(0) must haveHref(routes.RemoveItemsSummaryController.removeItem("id2"))
       }
 
       "item has two package information elements with one having empty number of packages" in {
@@ -151,7 +151,7 @@ class ItemsSummaryViewSpec extends UnitViewSpec with ExportsTestHelper with Stub
         rows.get(1).getElementById("item_0--item_type") must containText("1234567890")
         rows.get(1).getElementById("item_0--package_count").text() must be("1")
         rows.get(1).getElementById("item_0--change").getElementsByTag("a").get(0) must haveHref(routes.ProcedureCodesController.displayPage("id1"))
-        rows.get(1).getElementById("item_0--remove").getElementsByTag("a").get(0) must haveHref(routes.ItemsSummaryController.removeItem("id1"))
+        rows.get(1).getElementById("item_0--remove").getElementsByTag("a").get(0) must haveHref(routes.RemoveItemsSummaryController.removeItem("id1"))
 
       }
     }
